Fold plugin package install into operator
Remove services/plugins/store as a standalone package. Package installation is now an internal operator implementation detail:\n\nBefore:\n  import "github.com/valon-technologies/gestalt/server/services/plugins/store"\n  installed, err := store.Install(packagePath, destDir)\n\nAfter:\n  installed, err := installPackage(packagePath, destDir)\n\nThe unused InstallFromDir path and its unit tests are deleted. Source-directory installs already use the operator prepared-install flow, while archive installs remain covered through operator lifecycle and daemon E2E paths.
