+++ gestaltd/internal/operator/lifecycle.go
@@ -23,7 +23,6 @@ import (
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
 	pluginservice "github.com/valon-technologies/gestalt/server/services/plugins"
 	"github.com/valon-technologies/gestalt/server/services/plugins/providerpkg"
-	"github.com/valon-technologies/gestalt/server/services/plugins/store"
 	"gopkg.in/yaml.v3"
 )
 
@@ -1901,7 +1900,7 @@ func localUILockEntryFromPreparedInstall(paths lifecyclePaths, name string, plug
 	}, nil
 }
 
-func (l *Lifecycle) installMetadataSourcePackage(ctx context.Context, expectedKind, name, subject, destDir string, plugin *config.ProviderEntry, configDir string) (*store.InstalledPlugin, LockEntry, error) {
+func (l *Lifecycle) installMetadataSourcePackage(ctx context.Context, expectedKind, name, subject, destDir string, plugin *config.ProviderEntry, configDir string) (*installedPackage, LockEntry, error) {
 	sourceLocation := plugin.SourceReleaseLocation()
 	metadata, resolvedMetadataLocation, gitHubReleaseAssets, err := fetchProviderReleaseMetadata(ctx, l.metadataHTTPClient(), sourceLocation, sourceAuthToken(plugin))
 	if err != nil {
@@ -1942,7 +1941,7 @@ func (l *Lifecycle) installMetadataSourcePackage(ctx context.Context, expectedKi
 		return nil, LockEntry{}, fmt.Errorf("metadata source digest mismatch for %s: got %s, want %s", subject, download.SHA256Hex, archive.SHA256)
 	}
 
-	installed, err := store.Install(download.LocalPath, destDir)
+	installed, err := installPackage(download.LocalPath, destDir)
 	if err != nil {
 		return nil, LockEntry{}, fmt.Errorf("install metadata source package for %s: %w", subject, err)
 	}
@@ -2017,7 +2016,7 @@ func (l *Lifecycle) lockComponentEntryForSource(ctx context.Context, paths lifec
 
 	sourceLocation := plugin.SourceReleaseLocation()
 	var (
-		installed *store.InstalledPlugin
+		installed *installedPackage
 		entry     LockEntry
 		err       error
 	)
@@ -2073,7 +2072,7 @@ func (l *Lifecycle) lockProviderEntryForSource(ctx context.Context, paths lifecy
 	sourceLocation := plugin.SourceReleaseLocation()
 	destDir := providerDestDir(paths, name)
 	var (
-		installed *store.InstalledPlugin
+		installed *installedPackage
 		entry     LockEntry
 		err       error
 	)
@@ -2142,7 +2141,7 @@ func (l *Lifecycle) writeNamedUIProviderArtifact(ctx context.Context, paths life
 	expectedPackage := plugin.SourceReleaseLocation()
 
 	var (
-		installed *store.InstalledPlugin
+		installed *installedPackage
 		entry     LockEntry
 		opErr     error
 	)
@@ -2271,7 +2270,7 @@ func (l *Lifecycle) applyLockedProviders(configPaths []string, state StatePaths,
 	return validatedDuringPrepare, nil
 }
 
-func installLockedPackageAtomic(packagePath, destDir string) (*store.InstalledPlugin, func() error, func() error, error) {
+func installLockedPackageAtomic(packagePath, destDir string) (*installedPackage, func() error, func() error, error) {
 	if destDir == "" {
 		return nil, nil, nil, fmt.Errorf("destination directory is required")
 	}
@@ -2284,7 +2283,7 @@ func installLockedPackageAtomic(packagePath, destDir string) (*store.InstalledPl
 		return nil, nil, nil, fmt.Errorf("create temp install directory: %w", err)
 	}
 	cleanupDir := tempDir
-	installed, err := store.Install(packagePath, tempDir)
+	installed, err := installPackage(packagePath, tempDir)
 	if err != nil {
 		_ = os.RemoveAll(cleanupDir)
 		return nil, nil, nil, err
@@ -2998,7 +2997,7 @@ func (l *Lifecycle) materializeLockedUIProvider(ctx context.Context, paths lifec
 	if err := os.RemoveAll(destDir); err != nil {
 		return fmt.Errorf("remove stale cache for ui provider: %w", err)
 	}
-	installed, err := store.Install(download.LocalPath, destDir)
+	installed, err := installPackage(download.LocalPath, destDir)
 	if err != nil {
 		return fmt.Errorf("install locked source for ui provider: %w", err)
 	}
